fix sevral bugs

+++ cityflow_database/utils/processor.py
@@ -94,9 +94,10 @@ def delete_module(id):
     module = get_module(id)
     if module:
         icon = module.get('icon')
-        if icon.startswith('/api/dataset/source'):
-            icon_path = os.path.join(source_folder,'icons',os.path.basename(icon))
-            delete_file(icon_path)
+        if icon:
+            if icon.startswith('/api/dataset/source'):
+                icon_path = os.path.join(source_folder,'icons',os.path.basename(icon))
+                delete_file(icon_path)
         files = module.get('files',[])
         for file in files:
             if file['data'].startswith('/api/dataset/source'):
@@ -167,9 +168,10 @@ def delete_workflow(id):
     workflow = get_workflow(id)
     if workflow:
         screenshot = workflow.get('screenShot')
-        source_folder = os.getenv('DATABASE_SOURCE_DIR')
-        screenshot_path = os.path.join(source_folder,'images',os.path.basename(screenshot))
-        delete_file(screenshot_path)
+        if screenshot:
+            source_folder = os.getenv('DATABASE_SOURCE_DIR')
+            screenshot_path = os.path.join(source_folder,'images',os.path.basename(screenshot))
+            delete_file(screenshot_path)
         for node in workflow['nodes']:
             delete_module(node.get('id'))
     return delete_node('Workflow',id)
